Implemented multi-type variables.
Some node parameters can accept more than one data type.  For example, a
node parameter might accept either a Door or a Window as a value.  Multiple
types are specified as a comma-separated list of type names.

EditNodeVariableDialog was overhauled to display the referenced variable
if any, as well as display an error if the referenced node or variable doesn't
exist or the type of the referenced variable is incompatible.